requirements.txt
playwright==1.37.0
Pillow==10.0.0
streamlit==1.24.0
PyPDF2==3.0.1
pandas==2.0.3
numpy==1.24.3
random2==1.0.1 